dir structure cleanup
