+++ siem_query_utils/kql/Sentinel-ConnectorHealth.kql
