+++ siem_query_utils/kql/distinct-aadapps.kql
