+++ siem_query_utils/kql/distinct-aadidentities.kql
