+++ siem_query_utils/kql/distinct-endpointdevices.kql
