+++ siem_query_utils/kql/distinct-identities.kql
