+++ siem_query_utils/kql/distinct-tenantids.kql
