+++ siem_query_utils/kql/graph-workspaces.kql
